+++ android-auto-app/src/main/java/com/mapbox/maps/testapp/auto/car/CarLocationPuck.kt
@@ -1,11 +1,11 @@
 package com.mapbox.maps.testapp.auto.car
 
 import android.annotation.SuppressLint
-import androidx.car.app.CarContext
+import com.mapbox.maps.ImageHolder
 import com.mapbox.maps.R
-import com.mapbox.maps.extension.style.expressions.dsl.generated.literal
 import com.mapbox.maps.plugin.LocationPuck2D
 import com.mapbox.maps.plugin.LocationPuck3D
+import com.mapbox.maps.plugin.ModelScaleMode
 
 /**
  * Provides car location puck definitions.
@@ -14,18 +14,20 @@ internal object CarLocationPuck {
   /**
    * 3D location puck with the real world size.
    */
-  val duckLocationPuckHighZoom = LocationPuck3D(
+  val duckLocationPuckRealWorld = LocationPuck3D(
     modelUri = "https://raw.githubusercontent.com/KhronosGroup/glTF-Sample-Models/master/2.0/Duck/glTF-Embedded/Duck.gltf",
-    modelScaleExpression = literal(listOf(10, 10, 10)).toJson(),
+    modelScale = listOf(20f, 20f, 20f),
+    modelScaleMode = ModelScaleMode.MAP,
     modelRotation = listOf(0f, 0f, -90f)
   )
 
   /**
    * 3D location puck with a constant size across zoom levels.
    */
-  val duckLocationPuckLowZoom = LocationPuck3D(
+  val duckLocationPuckConstantSize = LocationPuck3D(
     modelUri = "https://raw.githubusercontent.com/KhronosGroup/glTF-Sample-Models/master/2.0/Duck/glTF-Embedded/Duck.gltf",
-    modelScale = listOf(0.2f, 0.2f, 0.2f),
+    modelScale = listOf(20f, 20f, 20f),
+    modelScaleMode = ModelScaleMode.VIEWPORT,
     modelRotation = listOf(0f, 0f, -90f)
   )
 
@@ -33,9 +35,9 @@ internal object CarLocationPuck {
    * Classic 2D location puck with blue dot and arrow.
    */
   @SuppressLint("UseCompatLoadingForDrawables")
-  fun classicLocationPuck2D(carContext: CarContext) = LocationPuck2D(
-    topImage = carContext.getDrawable(R.drawable.mapbox_user_icon),
-    bearingImage = carContext.getDrawable(R.drawable.mapbox_user_bearing_icon),
-    shadowImage = carContext.getDrawable(R.drawable.mapbox_user_stroke_icon)
+  val classicLocationPuck2D = LocationPuck2D(
+    topImage = ImageHolder.Companion.from(R.drawable.mapbox_user_icon),
+    bearingImage = ImageHolder.Companion.from(R.drawable.mapbox_user_bearing_icon),
+    shadowImage = ImageHolder.Companion.from(R.drawable.mapbox_user_stroke_icon)
   )
 }