+++ android-auto-app/src/main/java/com/mapbox/maps/testapp/auto/car/CarMapWidgets.kt
@@ -1,11 +1,11 @@
 package com.mapbox.maps.testapp.auto.car
 
+import com.mapbox.common.Cancelable
 import com.mapbox.maps.MapboxExperimental
 import com.mapbox.maps.extension.androidauto.MapboxCarMapObserver
 import com.mapbox.maps.extension.androidauto.MapboxCarMapSurface
 import com.mapbox.maps.extension.androidauto.widgets.CompassWidget
 import com.mapbox.maps.extension.androidauto.widgets.LogoWidget
-import com.mapbox.maps.plugin.delegates.listeners.OnCameraChangeListener
 
 /**
  * Note that the Widgets are only available when using android auto extension together with the
@@ -15,7 +15,7 @@ import com.mapbox.maps.plugin.delegates.listeners.OnCameraChangeListener
 class CarMapWidgets : MapboxCarMapObserver {
   private lateinit var logoWidget: LogoWidget
   private lateinit var compassWidget: CompassWidget
-  private lateinit var onCameraChangeListener: OnCameraChangeListener
+  private var cancellable: Cancelable? = null
   override fun onAttached(mapboxCarMapSurface: MapboxCarMapSurface) {
     super.onAttached(mapboxCarMapSurface)
     with(mapboxCarMapSurface) {
@@ -25,17 +25,19 @@ class CarMapWidgets : MapboxCarMapObserver {
         marginX = 26f,
         marginY = 120f,
       )
-      onCameraChangeListener = OnCameraChangeListener { compassWidget.setRotation(-mapSurface.getMapboxMap().cameraState.bearing.toFloat()) }
       mapSurface.addWidget(logoWidget)
       mapSurface.addWidget(compassWidget)
-      mapSurface.getMapboxMap().addOnCameraChangeListener(onCameraChangeListener)
+      cancellable = mapSurface.mapboxMap.subscribeCameraChanged {
+        compassWidget.setRotation(-mapSurface.mapboxMap.cameraState.bearing.toFloat())
+      }
     }
   }
 
   override fun onDetached(mapboxCarMapSurface: MapboxCarMapSurface) {
     super.onDetached(mapboxCarMapSurface)
     with(mapboxCarMapSurface) {
-      mapSurface.getMapboxMap().removeOnCameraChangeListener(onCameraChangeListener)
+      cancellable?.cancel()
+      cancellable = null
       mapSurface.removeWidget(logoWidget)
       mapSurface.removeWidget(compassWidget)
     }